Added custom error handler for 404 in __init__, added utils file for common functions
